--fix the some errors

+++ chiba-staking-contractV2/IPoolExtension.sol
@@ -10,6 +10,6 @@ interface IPoolExtension {
   ) external;
   
   function addTokenPool(
-    uint256 _addedAPR
+    uint8 _addedAPR
   ) external;
 }
